rename : common 폴더 경로를 components 폴더 안으로 변경
