+++ src/components/common/button/index.js
@@ -1,20 +1,3 @@
-{
-  /* <div class="button-container">
-<button
-  type="button"
-  class="button button--secondary text-caption"
-  id="cancel-button"
->
-  취소하기
-</button>
-<button
-  class="button button--primary text-caption"
-  id="register-button"
->
-  추가하기
-</button>
-</div> */
-}
 const Button = ({ text, style, onClick, type = "submit", id }) => {
   const button = document.createElement("button");
